new: packs calculation algo

+++ internal/domain/pack.go
@@ -1,5 +1,5 @@
 package domain
 
 type Pack struct {
-	Quantity int64 `json:"quantity"`
+	Quantity int `json:"quantity"`
 }
